Code review w/ Emmanuel. Most significant change: +src/edits/fakeAstEdits.js
