+++ src/nodeSpec.js
@@ -1,8 +1,5 @@
-import {warn, poscmp, srcRangeContains} from './utils';
-import {Blank, FakeInsertNode} from './nodes';
 import {ASTNode} from './ast';
 import hashObject from 'object-hash';
-
 
 
 // A NodeSpec declares the types of the fields of an ASTNode.
@@ -38,40 +35,6 @@ export function list(fieldName) {
 // value :: any -> ChildSpec
 export function value(fieldName) {
   return new Value(fieldName);
-}
-
-// Private-ish - used only in `edits/`
-export function findInsertionPoint(parent, fieldName, pos) {
-  return new InsertionPoint(parent, fieldName, pos);
-}
-
-// Private-ish - used only in `edits/`
-export function findReplacementPoint(parent, child) {
-  return new ReplacementPoint(parent, child);
-}
-
-// Private-ish - used only in `edits/`
-export function cloneNode(oldNode) {
-  let newNode = new ASTNode(oldNode.from, oldNode.to, oldNode.type, oldNode.options);
-  for (const spec of oldNode.spec.childSpecs) {
-    if (spec instanceof Required || spec instanceof Optional) {
-      if (oldNode[spec.fieldName]) {
-        newNode[spec.fieldName] = cloneNode(oldNode[spec.fieldName]);
-      } else {
-        newNode[spec.fieldName] = null;
-      }
-    } else if (spec instanceof Value) {
-      newNode[spec.fieldName] = oldNode[spec.fieldName];
-    } else if (spec instanceof List) {
-      newNode[spec.fieldName] = oldNode[spec.fieldName].map(cloneNode);
-    }
-  }
-  newNode.type = oldNode.type;
-  newNode.id = oldNode.id;
-  newNode.hash = oldNode.hash;
-  newNode.spec = oldNode.spec;
-  newNode.pretty = oldNode.pretty;
-  return newNode;
 }
 
 class NodeSpec {
@@ -150,108 +113,9 @@ class HashIterator {
   }
 }
 
-class InsertionPoint {
-  constructor(parent, fieldName, pos) {
-    this.parent = parent;
-    this.pos = pos;
-    // Find the spec that matches the supplied field name.
-    let spec = null;
-    for (const s of parent.spec.childSpecs) {
-      if (s instanceof List && s.fieldName === fieldName) {
-        spec = s;
-      }
-    }
-    if (spec === null) {
-      warn('NodeSpec', "Failed to find list to insert child into.");
-    }
-    this.spec = spec;
-    // If `pos` is null, that means the list is empty.
-    if (pos === null) {
-      this.index = 0;
-      return;
-    }
-    // `pos` lies inside the list. Find out where.
-    const list = parent[fieldName];
-    for (const i in list) {
-      if (poscmp(pos, list[i].srcRange().from) <= 0) {
-        this.index = i;
-        return;
-      }
-    }
-    this.index = list.length;
-    return;
-  }
-
-  insertChild(clonedParent, text) {
-    const newChildNode = new FakeInsertNode(this.pos, this.pos, text);
-    clonedParent[this.spec.fieldName].splice(this.index, 0, newChildNode);
-  }
-
-  findChild(newAST) {
-    const newParent = newAST.getNodeById(this.parent.id);
-    if (!newParent) return null;
-    return newParent[this.spec.fieldName][this.index];
-  }
-}
-
-class ReplacementPoint {
-  constructor(parent, child) {
-    this.parent = parent;
-    this.child = child;
-    for (const spec of parent.spec.childSpecs) {
-      const field = parent[spec.fieldName];
-      if (spec instanceof Required && field.id === child.id) {
-        this.spec = spec;
-        return;
-      } else if (spec instanceof Optional && field && field.id === child.id) {
-        this.spec = spec;
-        return;
-      } else if (spec instanceof List) {
-        for (const i in field) {
-          if (field[i].id === child.id) {
-            this.spec = spec;
-            this.index = i;
-            return;
-          }
-        }
-      }
-    }
-    warn('new ReplacementPoint', "Failed to find child to be replaced/deleted.");
-  }
-
-  replaceChild(clonedParent, text) {
-    const newChildNode = new FakeInsertNode(this.child.from, this.child.to, text);
-    if (this.index) {
-      clonedParent[this.spec.fieldName][this.index] = newChildNode;
-    } else {
-      clonedParent[this.spec.fieldName] = newChildNode;
-    }
-  }
-
-  deleteChild(clonedParent) {
-    if (this.index) {
-      clonedParent[this.spec.fieldName].splice(this.index, 1); // Remove the i'th element.
-    } else if (this.spec instanceof Optional) {
-      clonedParent[this.spec.fieldName] = null;
-    } else {
-      clonedParent[this.spec.fieldName] = new Blank(this.child.from, this.child.to);
-    }
-  }
-
-  findChild(newAST) {
-    const newParent = newAST.getNodeById(this.parent.id);
-    if (!newParent) return null;
-    if (this.index) {
-      return this.parent[this.spec.fieldName][this.index];
-    } else {
-      return this.parent[this.spec.fieldName];
-    }
-  }
-}
-
 class ChildSpec {}
 
-class Required extends ChildSpec {
+export class Required extends ChildSpec {
   constructor(fieldName) {
     super();
     this.fieldName = fieldName;
@@ -264,7 +128,7 @@ class Required extends ChildSpec {
   }
 }
 
-class Optional extends ChildSpec {
+export class Optional extends ChildSpec {
   constructor(fieldName) {
     super();
     this.fieldName = fieldName;
@@ -278,7 +142,7 @@ class Optional extends ChildSpec {
   }
 }
 
-class List extends ChildSpec {
+export class List extends ChildSpec {
   constructor(fieldName) {
     super();
     this.fieldName = fieldName;
@@ -300,7 +164,7 @@ class List extends ChildSpec {
   }
 }
 
-class Value extends ChildSpec {
+export class Value extends ChildSpec {
   constructor(fieldName) {
     super();
     this.fieldName = fieldName;
